RASD, Version 4.2
Addition of minor sections, such as references, hours of work, used
tools. Improved description of the class diagram. Started the
traceability table and the refining process of use cases (bigger use
cases now will be composed of multiple sub use cases). Such decision
was taken to a better realisation of the previously said table, and for
future usage.