RASD Release, Version 4.5
New folder structure. “Releases” folder will contain the final version
of documents, while “Presentations” folder will, as the name implies,
contain all the presentations regarding the documents of the project.
Every folder with the name of a document, “RASD” as example, will
contain all the materials needed to write the related document.